requirements.txt
requests==2.26.0
camelot-py==0.10.1
opencv-python==4.5.3.56
